requirements.txt
Flask>=2.0,<3.0
flask-mysqldb>=0.2.0
PyYAML>=5.4